Move header and content to new components
