+++ src/scenes/ModulesRouter/ListView.tsx
@@ -1,28 +1,20 @@
-import React, { useMemo } from 'react';
+import React from 'react';
 
 import { Box, Typography, useTheme } from '@mui/material';
-import { Dropdown, DropdownItem } from '@terraware/web-components';
 
 import Card from 'src/components/common/Card';
 import PageWithModuleTimeline from 'src/components/common/PageWithModuleTimeline';
-import useNavigateTo from 'src/hooks/useNavigateTo';
 import { useParticipantData } from 'src/providers/Participant/ParticipantContext';
-import { selectProjectModuleList } from 'src/redux/features/modules/modulesSelectors';
-import { useAppSelector } from 'src/redux/store';
 import strings from 'src/strings';
 
 import CurrentTimeline from './CurrentTimeline';
-import ModuleEntry from './ModuleEntry';
+import ListModulesContent from './ListModulesContent';
+import ListViewHeader from './ListViewHeader';
 
 export default function ListView(): JSX.Element {
   const theme = useTheme();
 
-  const { currentParticipant, currentParticipantProject, participantProjects, setCurrentParticipantProject } =
-    useParticipantData();
-
-  const { goToModules } = useNavigateTo();
-
-  const modules = useAppSelector(selectProjectModuleList(currentParticipantProject?.id ?? -1));
+  const { currentParticipant } = useParticipantData();
 
   // TODO - where will this be stored? Is this stored in the back end within another enum table?
   // Should we store it and localize it in the front end? Will it be stored somewhere an admin can edit it?
@@ -34,45 +26,10 @@ export default function ListView(): JSX.Element {
     'review is displayed in your To Do list on your home screen. Please login to Terraware regularly to check which ' +
     'deliverables are due or need review.';
 
-  const options: DropdownItem[] = useMemo(
-    () =>
-      participantProjects.map((project) => ({
-        label: project.name,
-        value: project.id,
-      })),
-    [participantProjects]
-  );
-
-  const selectStyles = {
-    arrow: {
-      height: '32px',
-    },
-    input: {
-      fontSize: '24px',
-      fontWeight: '600',
-      lineHeight: '32px',
-    },
-    inputContainer: {
-      border: 0,
-      backgroundColor: 'initial',
-    },
-  };
-
   return (
     <PageWithModuleTimeline title={strings.ALL_MODULES}>
       <Box sx={{ paddingBottom: 2 }}>
-        <Dropdown
-          onChange={(id) => {
-            const projectId = +id;
-            if (projectId != currentParticipantProject?.id) {
-              setCurrentParticipantProject(projectId);
-              goToModules(+projectId);
-            }
-          }}
-          options={options}
-          selectStyles={selectStyles}
-          selectedValue={currentParticipantProject?.id}
-        />
+        <ListViewHeader />
       </Box>
 
       <Card style={{ width: '100%' }}>
@@ -82,10 +39,7 @@ export default function ListView(): JSX.Element {
           <Typography>{phaseDescription}</Typography>
         </Box>
 
-        {currentParticipantProject &&
-          modules?.map((module, index) => (
-            <ModuleEntry index={index} key={index} module={module} projectId={currentParticipantProject?.id} />
-          ))}
+        <ListModulesContent />
       </Card>
     </PageWithModuleTimeline>
   );
